@@ -9,20 +9,32 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "vue": "^3.2.13",
+    "core-js": "^3.36.0",
+    "vue": "^3.4.0",
     "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
+    "@babel/core": "^7.23.0",
+    "@babel/eslint-parser": "^7.23.0",
+    "@vue/cli-plugin-babel": "^5.0.8",
     "@vue/cli-plugin-eslint": "^3.8.0",
-    "@vue/cli-service": "^3.3.1",
+    "@vue/cli-service": "^3.12.1",
     "@vue/eslint-config-prettier": "^10.2.0",
+    "case-sensitive-paths-webpack-plugin": "^2.4.0",
     "eslint": "^9.23.0",
-    "eslint-plugin-vue": "^9.33.0",
-    "gh-pages": "^6.3.0"
+    "eslint-plugin-vue": "^9.25.0",
+    "gh-pages": "^6.3.0",
+    "html-webpack-plugin": "^5.5.3"
+  },
+  "resolutions": {
+    "ejs": "^3.1.10",
+    "serialize-javascript": "^6.0.1",
+    "glob": "^9.0.0",
+    "rimraf": "^4.0.0"
+  },
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=8.0.0"
   },
   "eslintConfig": {
     "root": true,
@@ -44,9 +56,15 @@
     }
   },
   "browserslist": [
+    "defaults",
     "> 1%",
-    "last 2 versions",
     "not dead",
-    "not ie 11"
+    "not IE 11",
+    "maintained node versions",
+    "last 2 versions",
+    "last 2 Chrome versions",
+    "last 2 Firefox versions",
+    "last 2 Edge versions",
+    "last 2 Safari versions"
   ]
 }
