@@ -1,5 +1,11 @@
 module.exports = {
   presets: [
-    '@vue/cli-plugin-babel/preset'
+    [
+      '@vue/cli-plugin-babel/preset',
+      {
+        useBuiltIns: 'usage',
+        corejs: 3
+      }
+    ]
   ]
 }
